分けた&二重up通知は多分治った
